// packages/leeloo/src/team/types.ts
export enum TacTabs {
  ACCOUNT = "account",
  ACTIVITY = "activity",
  CHANGE_PLAN = "account/changeplan",
  DASHBOARD = "dashboard",
  GROUPS = "groups",
  MEMBERS = "members",
  SETTINGS = "settings",
  DARK_WEB_INSIGHTS = "dark-web-insights",
  INTEGRATIONS = "integrations",
  LOGGED_OUT_MONITORING = "logged-out-monitoring",
  POST_TRIAL_CHECKOUT = "account/checkout",
}
export type HeadingLevel = "h1" | "h2" | "h3" | "h4" | "h5" | "h6";